+++ CHANGELOG.md
@@ -1,14 +1,19 @@
+# 0.1.14
+
+* Add codemirror (and react-codemirror2) to list of available packages.
+
 # 0.1.13
 
 * Add autocomplete
 * Can get reference to state from URL.
 * Fix bug causing visrs to not update their srcloc correctly.
+* Can actually refer to cljs.* and clojure.* in require form.
 
 
 # 0.1.12
 
 * Begin changelog.
 * Switch from putting VISrs in iframes to directly in the DOM.
-* Add ability to call automatically call `main` from current module.
+* Add ability to call automatically `main` from current module.
 * Option to set tabl mode
 * defvisr indentation works as expected.